Add files via upload
This is a Online Banking System built using Java for the frontend and backend, Mysql for the database. In this project I tried to show the working of a banking account system and cover the basic functionality of a Bank Account Management System such as Login ,SignUp, Homepage, Deposite, Pin Generate, withdrarw.